LCOGT sched: tuning airmass for job acceptance
With my max airmass constraint on the astropy scheduler, I get a bunch of API
responses that I'm going too low. Even though MY maximum airmass is higher than
what they're setting.

SO.

For purposes of "OIBEO" transit submission, I'm limiting it to airmass<1.8
transits.

There are fewer of these, so the IBE and other partials will need to be
considered as well.